Disable visual tests (for now)
